Sandbox app and enable hardened runtime
In order to better follow the Apple Simple Firewall example code additional
security features are enabled. The main application is sandboxed with the
hardened runtime enabled. The system extension also has the hardened runtime enabled. The endpoint security extension can't be sandboxed because then it won't be able to communicate with the endpoint security subsystem.